Feature/436 address warnings (#537)
* Fix CA1033: Interface methods should be callable by child types

* Fix CA1064: Exceptions should be public

* Delete ADRException.cs

* Fix CA1508: avoid dead code in conditionals

* Fix CA1716: Do not use reserved keyword for members

* Fix CA1724: Resolve namespace conflicts

* Fix CA1805: Member initialized explicitly to default

* Fix CA1812: Remove unused classes

* Fix CA1822: Member does not access instance data and can be marked as static

* Fix IDE0078: Use pattern matching

* Fix IDE0120: Simplify LINQ expression

* Fix IDE0075: Conditional expression can be simplified

* Fix IDE0071: Interpolation can be simplified

* Fix IDE0063: Simplify using statement

* Fix IDE0059: Unnecessary assignment of a value

* Fix IDE0057: Slice can be simplified

* Fixe IDE0056: Indexing can be simplified

* Fix IDE0054: Use compound assignment

* Fix IDE0017: Object initialization can be simplified

* Fix IDE0004: Redundant cast

* Fix IDE0045: Simplify if-statement

* Fix IDE0041: Null check can be simplified

* Fix IDE0036: Modifiers are not ordered

* Fix IDE0034: Simplify default expression

* Fix IDE0029: Simplify null-check

* Fix xUnit1013: Dispose should be marked as Fact

* Fix IDE0027: Use expression body for accessors

* Fix IDE0024: Use expression body for operators

* Fix IDE0010: Populate switch statement

* Fix IDE0007: Use var instead of explicit type

* Fix CS1998: This async method lacks 'await'

* Fix CS0114: Inherited member hidden

* Fix CA2225: Provide named alternatives to operator overloads

* Fix CA2201: Do not raise reserved exception types

* Fix CA2200: Rethrowing caught exception changes stack information

* Fix CA2008: Pass TaskScheduler when creating tasks

* Fix CA2000: Disposable

* Fix CA1825: Avoid unnecessary zero-length array allocations

* Fix failing test

* Fix: CA2227: Collection properties should be read only

* Fix failing test after cherry-pick

* Fix inverted conditional that causes test failure

* merge issues

* fix IDE0003 and IDE0009 after merge

* Fix broken test

+++ LoRaEngine/modules/LoRaWanNetworkSrvModule/LoRaWan.NetworkServer/LoRaOperationTimeWatcher.cs
@@ -160,12 +160,12 @@ namespace LoRaWan.NetworkServer
 
         bool InTimeForJoinAcceptFirstWindow(TimeSpan elapsed)
         {
-            return elapsed.Add(ExpectedTimeToPackageAndSendMessage).TotalSeconds <= (double)this.loraRegion.JoinAcceptDelay1;
+            return elapsed.Add(ExpectedTimeToPackageAndSendMessage).TotalSeconds <= this.loraRegion.JoinAcceptDelay1;
         }
 
         bool InTimeForJoinAcceptSecondWindow(TimeSpan elapsed)
         {
-            return elapsed.Add(ExpectedTimeToPackageAndSendMessage).TotalSeconds <= (double)this.loraRegion.JoinAcceptDelay2;
+            return elapsed.Add(ExpectedTimeToPackageAndSendMessage).TotalSeconds <= this.loraRegion.JoinAcceptDelay2;
         }
 
         /// <summary>
